requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.2.0
backcall==0.2.0
colorama==0.4.3
decorator==4.4.2
Django==3.1.1
django-countries==6.1.3
django-tsvector-field==0.9.5
flake8==3.8.3
iniconfig==1.0.1
jedi==0.17.2
mccabe==0.6.1
packaging==20.4
parso==0.7.1
pickleshare==0.7.5
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==3.0.7
psycopg2-binary==2.8.4
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
pyparsing==2.4.7
pytest==6.1.1
pytest-django==3.10.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wcwidth==0.2.5
